fix #604
