+++ MLS.Agent.Tools/TypeExtensions.cs
@@ -35,12 +35,5 @@ namespace MLS.Agent.Tools
                 return reader.ReadToEnd();
             }
         }
-
-        public static bool CanBeInstantiated(this Type type)
-        {
-            return !type.IsAbstract
-                    && !type.IsGenericTypeDefinition
-                    && !type.IsInterface;
-        }
     }
 }
