+++ Microsoft.DotNet.Interactive.Rendering/FormatterSetBase.cs
@@ -12,17 +12,17 @@ namespace Microsoft.DotNet.Interactive.Rendering
             ConcurrentDictionary<Type, Func<Type, ITypeFormatter>> openGenericFormatterFactories = null,
             ConcurrentDictionary<Type, ITypeFormatter> formatters = null)
         {
-            OpenGenericFormatterFactories = openGenericFormatterFactories ??
+            FormatterFactories = openGenericFormatterFactories ??
                                             new ConcurrentDictionary<Type, Func<Type, ITypeFormatter>>();
             Formatters = formatters ??
                          new ConcurrentDictionary<Type, ITypeFormatter>();
         }
 
-        protected ConcurrentDictionary<Type, Func<Type, ITypeFormatter>> OpenGenericFormatterFactories { get; }
+        protected ConcurrentDictionary<Type, Func<Type, ITypeFormatter>> FormatterFactories { get; }
 
         protected ConcurrentDictionary<Type, ITypeFormatter> Formatters { get; }
 
-        public void AddFormatterFactoryForOpenGenericType(
+        public void AddFormatterFactory(
             Type type,
             Func<Type, ITypeFormatter> getFormatter)
         {
@@ -31,12 +31,7 @@ namespace Microsoft.DotNet.Interactive.Rendering
                 throw new ArgumentNullException(nameof(type));
             }
 
-            if (!type.IsGenericTypeDefinition)
-            {
-                throw new ArgumentException($"Type {type} is not an open generic type.");
-            }
-
-            OpenGenericFormatterFactories[type] = getFormatter;
+            FormatterFactories[type] = getFormatter;
         }
 
         public bool TryGetFormatterForType(Type type, out ITypeFormatter formatter)
@@ -44,7 +39,7 @@ namespace Microsoft.DotNet.Interactive.Rendering
             if (!Formatters.TryGetValue(type, out formatter))
             {
                 if (type.IsGenericType &&
-                    OpenGenericFormatterFactories.TryGetValue(
+                    FormatterFactories.TryGetValue(
                         type.GetGenericTypeDefinition(),
                         out var factory))
                 {
