+++ Microsoft.DotNet.Interactive.Rendering/Microsoft.DotNet.Interactive.Rendering.csproj
@@ -16,6 +16,10 @@
 
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="12.0.02" />
+    <PackageReference Include="pocket.disposable" Version="1.0.5">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
     <PackageReference Include="System.Memory" Version="4.5.3" />
   </ItemGroup>
