+++ Microsoft.DotNet.Interactive.Rendering/NonDestructurer.cs
@@ -1,0 +1,20 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Collections.Generic;
+
+namespace Microsoft.DotNet.Interactive.Rendering
+{
+    internal class NonDestructurer : IDestructurer
+    {
+        public static IDestructurer Instance { get; } = new NonDestructurer();
+
+        public IDictionary<string, object> Destructure(object instance)
+        {
+            return new Dictionary<string, object>
+            {
+                ["value"] = instance
+            };
+        }
+    }
+}
